refactor: consolidate callback methods back into callbacks.go
The split into callbacks_search.go and callbacks_changes.go added
file-hopping without adding clarity — the methods are only referenced
from registerWidgetCallbacks and belong together. The named methods
stay extracted (not re-inlined), just colocated in one file.
