+++ cmd/ttt/callbacks.go
@@ -1,7 +1,13 @@
 package main
 
 import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+
 	"github.com/eugenioenko/ttt/internal/command"
+	"github.com/eugenioenko/ttt/internal/core/diff"
 	"github.com/eugenioenko/ttt/internal/git"
 	"github.com/eugenioenko/ttt/internal/ui"
 
@@ -40,6 +46,251 @@ func (a *App) ShowSidebarMoreMenu(sx, sy int) {
 	if len(items) > 0 {
 		openContextMenu(a, items, sx, sy)
 	}
+}
+
+func (a *App) DiffSearchSources() []ui.DiffSearchSource {
+	seen := map[string]bool{}
+	sources := a.editorGroup.DiffTabSources()
+	for _, s := range sources {
+		seen[s.TabName] = true
+	}
+	for _, g := range a.changes.Groups {
+		if !g.IsPR {
+			continue
+		}
+		for path, diffText := range g.PRDiffs {
+			tabName := path + " (diff)"
+			if seen[tabName] {
+				continue
+			}
+			fd := diff.Parse(diffText)
+			dv := ui.NewDiffViewWidget(path, fd)
+			sources = append(sources, ui.DiffSearchSource{TabName: tabName, Lines: dv.CombinedLines()})
+		}
+	}
+	return sources
+}
+
+func (a *App) NavigateToSearchMatch(path string, line, col int) {
+	if strings.HasSuffix(path, " (diff)") {
+		if !a.editorGroup.SwitchToTabByPath(path) {
+			filePath := strings.TrimSuffix(path, " (diff)")
+			for _, g := range a.changes.Groups {
+				if !g.IsPR {
+					continue
+				}
+				if diffText, ok := g.PRDiffs[filePath]; ok {
+					a.editorGroup.OpenDiff(filePath, diff.Parse(diffText))
+					break
+				}
+			}
+		}
+		if dv := a.editorGroup.ActiveDiffWidget(); dv != nil {
+			dv.ScrollToLine(line - 1)
+			dv.ApplySearchHighlight(a.search.Input.Text, a.search.Options)
+		}
+		a.root.SetFocus(a.editorGroup)
+		return
+	}
+	a.editorGroup.OpenFile(path)
+	a.editorGroup.GoToLine(line)
+	if a.search.Input.Text != "" {
+		matches, _ := ui.FindInLines(a.editorGroup.Editor.Buf.Lines, a.search.Input.Text, a.search.Options)
+		a.editorGroup.SetSearch(a.search.Input.Text, matches)
+	}
+	a.root.SetFocus(a.editorGroup)
+}
+
+func (a *App) PreviewSearchReplace(filePath string, matches []ui.SearchMatch, replacement string, opts ui.SearchOptions) {
+	data, err := os.ReadFile(filePath)
+	if err != nil {
+		a.StatusWarn("Cannot read file: " + err.Error())
+		return
+	}
+	lines := strings.Split(string(data), "\n")
+	if len(lines) > 0 && lines[len(lines)-1] == "" {
+		lines = lines[:len(lines)-1]
+	}
+	fd := ui.BuildReplaceDiff(filepath.Base(filePath), lines, matches, replacement, opts)
+	a.editorGroup.OpenDiff(filePath, fd)
+	a.root.SetFocus(a.editorGroup)
+}
+
+func (a *App) ApplySearchReplace(filePath string, matches []ui.SearchMatch, replacement string, opts ui.SearchOptions) {
+	data, err := os.ReadFile(filePath)
+	if err != nil {
+		a.StatusWarn("Cannot read file: " + err.Error())
+		return
+	}
+	lines := strings.Split(string(data), "\n")
+	if len(lines) > 0 && lines[len(lines)-1] == "" {
+		lines = lines[:len(lines)-1]
+	}
+	newLines := ui.ApplyReplacements(lines, matches, replacement, opts)
+	if err := os.WriteFile(filePath, []byte(strings.Join(newLines, "\n")+"\n"), 0644); err != nil {
+		a.StatusWarn("Cannot write file: " + err.Error())
+		return
+	}
+	a.editorGroup.ReloadFile(filePath)
+	a.search.Refresh()
+	a.StatusNotify(fmt.Sprintf("Replaced %d matches in %s", len(matches), filepath.Base(filePath)))
+}
+
+func (a *App) ApplySearchReplaceAll(allMatches map[string][]ui.SearchMatch, replacement string, opts ui.SearchOptions) {
+	totalFiles := len(allMatches)
+	totalMatches := 0
+	for _, m := range allMatches {
+		totalMatches += len(m)
+	}
+	msg := fmt.Sprintf("Replace %d matches across %d files? This cannot be undone.", totalMatches, totalFiles)
+	a.ShowConfirmDialog(msg, []string{"Cancel", "Replace All"}, []func(){
+		func() { a.DismissDialog() },
+		func() {
+			a.DismissDialog()
+			for filePath, matches := range allMatches {
+				data, err := os.ReadFile(filePath)
+				if err != nil {
+					continue
+				}
+				lines := strings.Split(string(data), "\n")
+				if len(lines) > 0 && lines[len(lines)-1] == "" {
+					lines = lines[:len(lines)-1]
+				}
+				newLines := ui.ApplyReplacements(lines, matches, replacement, opts)
+				if err := os.WriteFile(filePath, []byte(strings.Join(newLines, "\n")+"\n"), 0644); err != nil {
+					continue
+				}
+				a.editorGroup.ReloadFile(filePath)
+			}
+			a.search.Refresh()
+			a.StatusNotify(fmt.Sprintf("Replaced %d matches across %d files", totalMatches, totalFiles))
+		},
+	})
+}
+
+func (a *App) OpenChangeDiff(dir string, status git.FileStatus) {
+	fullPath := filepath.Join(dir, status.Path)
+	if status.Status == "?" {
+		a.editorGroup.OpenFile(fullPath)
+		a.root.SetFocus(a.editorGroup)
+		return
+	}
+	var diffText string
+	var err error
+	if status.Status == "R" && status.OldPath != "" {
+		diffText, err = git.DiffRename(dir, status.OldPath, status.Path)
+	} else {
+		diffText, err = git.DiffFile(dir, status.Path)
+	}
+	if err != nil || diffText == "" {
+		a.editorGroup.OpenFile(fullPath)
+		a.root.SetFocus(a.editorGroup)
+		return
+	}
+	parsed := diff.Parse(diffText)
+	if len(parsed.Hunks) == 0 {
+		a.editorGroup.OpenFile(fullPath)
+		a.root.SetFocus(a.editorGroup)
+		return
+	}
+	a.editorGroup.OpenDiff(status.Path, parsed)
+	a.root.SetFocus(a.editorGroup)
+}
+
+func (a *App) OpenPRDiff(group *ui.ChangesGroup, status git.FileStatus) {
+	diffText, ok := group.PRDiffs[status.Path]
+	if !ok || diffText == "" {
+		a.StatusWarn("No diff available for " + status.Path)
+		return
+	}
+	parsed := diff.Parse(diffText)
+	if len(parsed.Hunks) == 0 {
+		a.StatusWarn("Empty diff for " + status.Path)
+		return
+	}
+	a.editorGroup.OpenDiff(status.Path, parsed)
+	a.root.SetFocus(a.editorGroup)
+}
+
+func (a *App) ShowPRGroupMenu(group *ui.ChangesGroup, sx, sy int) {
+	reg := a.reg
+	name := group.Name
+	url := group.PRURL
+	refreshID := "pr.refresh." + name
+	closeID := "pr.close." + name
+	reg.Register(command.Command{
+		ID: refreshID, Title: "Refresh",
+		Handler: func() {
+			a.changes.RemovePRGroup(name)
+			a.fetchAndOpenPR(url)
+		},
+	})
+	reg.Register(command.Command{
+		ID: closeID, Title: "Close",
+		Handler: func() {
+			a.changes.RemovePRGroup(name)
+		},
+	})
+	items := []ui.ContextMenuItem{
+		{Label: "Refresh", Command: refreshID},
+		{Label: "Close", Command: closeID},
+	}
+	openContextMenu(a, items, sx, sy)
+}
+
+func (a *App) ShowGroupMenu(dir string, sx, sy int) {
+	reg := a.reg
+	items := []ui.ContextMenuItem{
+		{Label: "Pull", Command: "git.pull." + dir},
+		{Label: "Push", Command: "git.push." + dir},
+		{Label: "Sync", Command: "git.sync." + dir},
+	}
+	registerDirGitCmd := func(id, title string, ops []func(string) error, verb string) {
+		reg.Register(command.Command{
+			ID: id, Title: title,
+			Handler: func() {
+				for _, op := range ops {
+					if err := op(dir); err != nil {
+						a.StatusError(fmt.Sprintf("%s failed: %v", verb, err))
+						return
+					}
+				}
+				a.StatusNotify(verb + " successfully")
+				a.changes.Refresh()
+			},
+		})
+	}
+	registerDirGitCmd("git.pull."+dir, "Pull", []func(string) error{git.Pull}, "Pulled")
+	registerDirGitCmd("git.push."+dir, "Push", []func(string) error{git.Push}, "Pushed")
+	registerDirGitCmd("git.sync."+dir, "Sync", []func(string) error{git.Pull, git.Push}, "Synced")
+	openContextMenu(a, items, sx, sy)
+}
+
+func (a *App) CommitChanges(dir string, message string) {
+	if err := git.Commit(dir, message); err != nil {
+		a.StatusError("Commit failed: " + err.Error())
+	} else {
+		for i := range a.changes.Groups {
+			if a.changes.Groups[i].Dir == dir {
+				a.changes.Groups[i].Input.Clear()
+			}
+		}
+		a.StatusNotify("Committed: " + message)
+		a.changes.Refresh()
+	}
+}
+
+func (a *App) ConfirmDiscard(message string, onConfirm func()) {
+	a.ShowConfirmDialog(message,
+		[]string{"Cancel", "Discard"},
+		[]func(){
+			func() { a.DismissDialog() },
+			func() {
+				a.DismissDialog()
+				onConfirm()
+			},
+		},
+	)
 }
 
 func registerWidgetCallbacks(app *App) {
