--- cmd/ttt/callbacks_search.go
@@ -1,131 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-	"strings"
-
-	"github.com/eugenioenko/ttt/internal/core/diff"
-	"github.com/eugenioenko/ttt/internal/ui"
-)
-
-func (a *App) DiffSearchSources() []ui.DiffSearchSource {
-	seen := map[string]bool{}
-	sources := a.editorGroup.DiffTabSources()
-	for _, s := range sources {
-		seen[s.TabName] = true
-	}
-	for _, g := range a.changes.Groups {
-		if !g.IsPR {
-			continue
-		}
-		for path, diffText := range g.PRDiffs {
-			tabName := path + " (diff)"
-			if seen[tabName] {
-				continue
-			}
-			fd := diff.Parse(diffText)
-			dv := ui.NewDiffViewWidget(path, fd)
-			sources = append(sources, ui.DiffSearchSource{TabName: tabName, Lines: dv.CombinedLines()})
-		}
-	}
-	return sources
-}
-
-func (a *App) NavigateToSearchMatch(path string, line, col int) {
-	if strings.HasSuffix(path, " (diff)") {
-		if !a.editorGroup.SwitchToTabByPath(path) {
-			filePath := strings.TrimSuffix(path, " (diff)")
-			for _, g := range a.changes.Groups {
-				if !g.IsPR {
-					continue
-				}
-				if diffText, ok := g.PRDiffs[filePath]; ok {
-					a.editorGroup.OpenDiff(filePath, diff.Parse(diffText))
-					break
-				}
-			}
-		}
-		if dv := a.editorGroup.ActiveDiffWidget(); dv != nil {
-			dv.ScrollToLine(line - 1)
-			dv.ApplySearchHighlight(a.search.Input.Text, a.search.Options)
-		}
-		a.root.SetFocus(a.editorGroup)
-		return
-	}
-	a.editorGroup.OpenFile(path)
-	a.editorGroup.GoToLine(line)
-	if a.search.Input.Text != "" {
-		matches, _ := ui.FindInLines(a.editorGroup.Editor.Buf.Lines, a.search.Input.Text, a.search.Options)
-		a.editorGroup.SetSearch(a.search.Input.Text, matches)
-	}
-	a.root.SetFocus(a.editorGroup)
-}
-
-func (a *App) PreviewSearchReplace(filePath string, matches []ui.SearchMatch, replacement string, opts ui.SearchOptions) {
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		a.StatusWarn("Cannot read file: " + err.Error())
-		return
-	}
-	lines := strings.Split(string(data), "\n")
-	if len(lines) > 0 && lines[len(lines)-1] == "" {
-		lines = lines[:len(lines)-1]
-	}
-	fd := ui.BuildReplaceDiff(filepath.Base(filePath), lines, matches, replacement, opts)
-	a.editorGroup.OpenDiff(filePath, fd)
-	a.root.SetFocus(a.editorGroup)
-}
-
-func (a *App) ApplySearchReplace(filePath string, matches []ui.SearchMatch, replacement string, opts ui.SearchOptions) {
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		a.StatusWarn("Cannot read file: " + err.Error())
-		return
-	}
-	lines := strings.Split(string(data), "\n")
-	if len(lines) > 0 && lines[len(lines)-1] == "" {
-		lines = lines[:len(lines)-1]
-	}
-	newLines := ui.ApplyReplacements(lines, matches, replacement, opts)
-	if err := os.WriteFile(filePath, []byte(strings.Join(newLines, "\n")+"\n"), 0644); err != nil {
-		a.StatusWarn("Cannot write file: " + err.Error())
-		return
-	}
-	a.editorGroup.ReloadFile(filePath)
-	a.search.Refresh()
-	a.StatusNotify(fmt.Sprintf("Replaced %d matches in %s", len(matches), filepath.Base(filePath)))
-}
-
-func (a *App) ApplySearchReplaceAll(allMatches map[string][]ui.SearchMatch, replacement string, opts ui.SearchOptions) {
-	totalFiles := len(allMatches)
-	totalMatches := 0
-	for _, m := range allMatches {
-		totalMatches += len(m)
-	}
-	msg := fmt.Sprintf("Replace %d matches across %d files? This cannot be undone.", totalMatches, totalFiles)
-	a.ShowConfirmDialog(msg, []string{"Cancel", "Replace All"}, []func(){
-		func() { a.DismissDialog() },
-		func() {
-			a.DismissDialog()
-			for filePath, matches := range allMatches {
-				data, err := os.ReadFile(filePath)
-				if err != nil {
-					continue
-				}
-				lines := strings.Split(string(data), "\n")
-				if len(lines) > 0 && lines[len(lines)-1] == "" {
-					lines = lines[:len(lines)-1]
-				}
-				newLines := ui.ApplyReplacements(lines, matches, replacement, opts)
-				if err := os.WriteFile(filePath, []byte(strings.Join(newLines, "\n")+"\n"), 0644); err != nil {
-					continue
-				}
-				a.editorGroup.ReloadFile(filePath)
-			}
-			a.search.Refresh()
-			a.StatusNotify(fmt.Sprintf("Replaced %d matches across %d files", totalMatches, totalFiles))
-		},
-	})
-}
